+++ packages/kbn-babel-preset/common_preset.js
