+++ packages/kbn-babel-preset/webpack_preset.js
@@ -13,6 +13,6 @@ module.exports = {
         useBuiltIns: true,
       },
     ],
-    require('./common'),
+    require('./common_preset'),
   ]
 };
