+++ packages/kbn-test/.babelrc
@@ -1,3 +1,3 @@
 {
-  "presets": ["@kbn/babel-preset/node"]
+  "presets": ["@kbn/babel-preset/node_preset"]
 }
