+++ packages/kbn-ui-framework/doc_site/webpack.config.js
@@ -27,7 +27,7 @@ module.exports = {
       exclude: /node_modules/,
       query: {
         presets: [
-          require.resolve('@kbn/babel-preset/webpack')
+          require.resolve('@kbn/babel-preset/webpack_preset')
         ],
       },
     }, {
