+++ src/dev/jest/babel_transform.js
@@ -2,6 +2,6 @@ const babelJest = require('babel-jest');
 
 module.exports = babelJest.createTransformer({
   presets: [
-    require.resolve('@kbn/babel-preset/node')
+    require.resolve('@kbn/babel-preset/node_preset')
   ]
 });
