+++ src/optimize/base_optimizer.js
@@ -13,7 +13,7 @@ import { fromRoot } from '../utils';
 import { PUBLIC_PATH_PLACEHOLDER } from './public_path_placeholder';
 
 const POSTCSS_CONFIG_PATH = require.resolve('./postcss.config');
-const BABEL_PRESET_PATH = require.resolve('@kbn/babel-preset/webpack');
+const BABEL_PRESET_PATH = require.resolve('@kbn/babel-preset/webpack_preset');
 const BABEL_EXCLUDE_RE = [
   /[\/\\](webpackShims|node_modules|bower_components)[\/\\]/,
 ];
